Breaking change, but you know... fixing the hash.
Hash is now hmac with the message type. The same payload for two messages
should be unique to the message type. I'm not totally abusing the hmac,
but maybe a little.

Also switched so the hash is stored as a long. The lookup will be faster
though unsure by how much. Greater than zero...